+++ brian-backend/Base/shared/base/InfoEnums.ts
@@ -41,6 +41,7 @@ export enum CollectionSource {
   KEYWORD = 'KEYWORD',
   RANDOM = 'RANDOM',
   CUSTOM = 'CUSTOM',
+  CURRENT = 'CURRENT',
 }
 
 /**