+++ shared/src/types/index.ts
@@ -38,6 +38,7 @@ export enum CollectionSource {
   KEYWORD = 'KEYWORD',
   RANDOM = 'RANDOM',
   CUSTOM = 'CUSTOM',
+  CURRENT = 'CURRENT',
 }
 
 export const ContextSource = CollectionSource;